notebooklm-api/pyproject.toml
[project]
name = "notebooklm-api"
version = "0.1.0"
description = "REST API backend wrapping notebooklm-py for transcript-to-video and all artifact generation"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "notebooklm-py==0.3.4",
    "aiofiles>=24.0.0",
    "playwright>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["app", "tests"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"app/routers/*.py" = ["B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
